refactor: split env_store, export, commands, and path_manage into focused modules
env_store.rs and path_manage.rs each split their Windows registry
implementation into a sibling *_backend.rs, mirroring the existing
EnvBackend trait abstraction. export.rs and commands.rs split into
directories with one module per format/domain (json, reg, ps1, dsc,
ansible / env, snapshot, export, path, launch).
