--- src-tauri/src/export.rs
@@ -1,768 +1,0 @@
-use std::collections::HashMap;
-
-use crate::env_store::{EnvSnapshot, EnvValue, EnvValueKind};
-use crate::error::EnvarlyError;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ExportScope {
-    User,
-    System,
-    All,
-}
-
-impl ExportScope {
-    pub fn includes_user(self) -> bool {
-        matches!(self, ExportScope::User | ExportScope::All)
-    }
-    pub fn includes_system(self) -> bool {
-        matches!(self, ExportScope::System | ExportScope::All)
-    }
-}
-
-impl TryFrom<&str> for ExportScope {
-    type Error = EnvarlyError;
-    fn try_from(s: &str) -> Result<Self, Self::Error> {
-        match s {
-            "User" => Ok(ExportScope::User),
-            "System" => Ok(ExportScope::System),
-            "All" => Ok(ExportScope::All),
-            other => Err(EnvarlyError::InvalidInput(format!(
-                "invalid scope: {other:?}"
-            ))),
-        }
-    }
-}
-
-pub struct FormatSpec {
-    pub content: String,
-    pub ext: &'static str,
-    pub file_stem: String,
-    pub filter_name: &'static str,
-}
-
-/// Resolve format string and scope to a ready-to-save export. `stem_prefix` is prepended
-/// to the format-specific filename suffix (e.g. "envarly" or "envarly-custom").
-pub fn resolve_export(
-    snapshot: &EnvSnapshot,
-    format: &str,
-    scope: ExportScope,
-    stem_prefix: &str,
-) -> FormatSpec {
-    let (content, ext, stem_suffix, filter_name) = match format {
-        "reg" => (to_reg(snapshot, scope), "reg", "", "Registry files"),
-        "ps1" => (to_ps1(snapshot, scope), "ps1", "", "PowerShell Script"),
-        "dsc_v2" => (
-            to_dsc_v2(snapshot, scope),
-            "ps1",
-            "-dsc",
-            "PowerShell DSC Configuration",
-        ),
-        "dsc_v3" => (
-            to_dsc_v3(snapshot, scope),
-            "dsc.yaml",
-            "-dsc3",
-            "DSC v3 Configuration",
-        ),
-        "ansible" => (
-            to_ansible(snapshot, scope),
-            "yml",
-            "-ansible",
-            "Ansible Playbook",
-        ),
-        _ => (to_json(snapshot, scope), "json", "", "JSON files"),
-    };
-    FormatSpec {
-        content,
-        ext,
-        file_stem: format!("{stem_prefix}{stem_suffix}"),
-        filter_name,
-    }
-}
-
-/// Export snapshot as pretty-printed JSON.
-pub fn to_json(snapshot: &EnvSnapshot, scope: ExportScope) -> String {
-    let mut obj = serde_json::Map::new();
-    obj.insert("version".into(), serde_json::Value::from(2));
-    if scope.includes_user() {
-        obj.insert(
-            "user".into(),
-            serde_json::to_value(&snapshot.user).unwrap_or_default(),
-        );
-    }
-    if scope.includes_system() {
-        obj.insert(
-            "system".into(),
-            serde_json::to_value(&snapshot.system).unwrap_or_default(),
-        );
-    }
-    serde_json::to_string_pretty(&serde_json::Value::Object(obj)).unwrap_or_default()
-}
-
-/// Export snapshot in Windows Registry .reg format.
-pub fn to_reg(snapshot: &EnvSnapshot, scope: ExportScope) -> String {
-    let mut out = String::from("Windows Registry Editor Version 5.00\r\n");
-    if scope.includes_user() {
-        out.push_str("\r\n[HKEY_CURRENT_USER\\Environment]\r\n");
-        let mut pairs: Vec<_> = snapshot.user.iter().collect();
-        pairs.sort_by_key(|(k, _)| k.to_lowercase());
-        for (k, v) in pairs {
-            out.push_str(&reg_entry(k, v));
-        }
-    }
-    if scope.includes_system() {
-        out.push_str(
-            "\r\n[HKEY_LOCAL_MACHINE\\SYSTEM\\CurrentControlSet\\Control\\Session Manager\\Environment]\r\n",
-        );
-        let mut pairs: Vec<_> = snapshot.system.iter().collect();
-        pairs.sort_by_key(|(k, _)| k.to_lowercase());
-        for (k, v) in pairs {
-            out.push_str(&reg_entry(k, v));
-        }
-    }
-    out
-}
-
-// ── PowerShell script ─────────────────────────────────────────────────────
-
-/// Export snapshot as a PowerShell script that preserves registry value types.
-pub fn to_ps1(snapshot: &EnvSnapshot, scope: ExportScope) -> String {
-    let mut out = String::from("# Generated by Envarly\n");
-    if scope.includes_user() && !snapshot.user.is_empty() {
-        out.push_str(
-            "$userEnvironment = [Microsoft.Win32.Registry]::CurrentUser.OpenSubKey('Environment', $true)\n",
-        );
-        out.push_str("\n# User environment variables\n");
-        let mut pairs: Vec<_> = snapshot.user.iter().collect();
-        pairs.sort_by_key(|(k, _)| k.to_lowercase());
-        for (k, v) in pairs {
-            out.push_str(&format!(
-                "$userEnvironment.SetValue({}, {}, {})\n",
-                ps1_quote(k),
-                ps1_quote(&v.value),
-                ps1_kind(v.kind),
-            ));
-        }
-        out.push_str("$userEnvironment.Dispose()\n");
-    }
-    if scope.includes_system() && !snapshot.system.is_empty() {
-        out.push_str(
-            "$systemEnvironment = [Microsoft.Win32.Registry]::LocalMachine.OpenSubKey('SYSTEM\\CurrentControlSet\\Control\\Session Manager\\Environment', $true)\n",
-        );
-        out.push_str("\n# System environment variables\n");
-        let mut pairs: Vec<_> = snapshot.system.iter().collect();
-        pairs.sort_by_key(|(k, _)| k.to_lowercase());
-        for (k, v) in pairs {
-            out.push_str(&format!(
-                "$systemEnvironment.SetValue({}, {}, {})\n",
-                ps1_quote(k),
-                ps1_quote(&v.value),
-                ps1_kind(v.kind),
-            ));
-        }
-        out.push_str("$systemEnvironment.Dispose()\n");
-    }
-    out
-}
-
-// ── PowerShell DSC v2 ─────────────────────────────────────────────────────
-
-/// Export snapshot as a PowerShell DSC v2 Configuration (.ps1).
-pub fn to_dsc_v2(snapshot: &EnvSnapshot, scope: ExportScope) -> String {
-    let mut resources = String::new();
-    if scope.includes_user() {
-        let mut pairs: Vec<_> = snapshot.user.iter().collect();
-        pairs.sort_by_key(|(k, _)| k.to_lowercase());
-        for (k, v) in pairs {
-            resources.push_str(&format!(
-                "        Environment {} {{\n            Name   = {}\n            Value  = {}\n            Target = @('User')\n        }}\n",
-                dsc_id(k, "User"),
-                ps1_quote(k),
-                ps1_quote(&v.value),
-            ));
-        }
-    }
-    if scope.includes_system() {
-        let mut pairs: Vec<_> = snapshot.system.iter().collect();
-        pairs.sort_by_key(|(k, _)| k.to_lowercase());
-        for (k, v) in pairs {
-            resources.push_str(&format!(
-                "        Environment {} {{\n            Name   = {}\n            Value  = {}\n            Target = @('Machine')\n        }}\n",
-                dsc_id(k, "Machine"),
-                ps1_quote(k),
-                ps1_quote(&v.value),
-            ));
-        }
-    }
-    format!(
-        "# Generated by Envarly\n{}Configuration EnvarlyConfig {{\n    Import-DscResource -ModuleName PSDscResources\n\n    Node localhost {{\n{}    }}\n}}\nEnvarlyConfig\n",
-        type_limitation_notice(snapshot, scope, "# "),
-        resources,
-    )
-}
-
-// ── PowerShell DSC v3 ─────────────────────────────────────────────────────
-
-/// Export snapshot as a DSC v3 config document (.dsc.yaml).
-pub fn to_dsc_v3(snapshot: &EnvSnapshot, scope: ExportScope) -> String {
-    let mut resources = String::new();
-    if scope.includes_user() {
-        let mut pairs: Vec<_> = snapshot.user.iter().collect();
-        pairs.sort_by_key(|(k, _)| k.to_lowercase());
-        for (k, v) in pairs {
-            resources.push_str(&format!(
-                "  - name: Set {} (User)\n    type: Microsoft.Windows/Environment\n    properties:\n      name: {}\n      value: {}\n      scope: CurrentUser\n",
-                k,
-                yaml_quote(k),
-                yaml_quote(&v.value),
-            ));
-        }
-    }
-    if scope.includes_system() {
-        let mut pairs: Vec<_> = snapshot.system.iter().collect();
-        pairs.sort_by_key(|(k, _)| k.to_lowercase());
-        for (k, v) in pairs {
-            resources.push_str(&format!(
-                "  - name: Set {} (System)\n    type: Microsoft.Windows/Environment\n    properties:\n      name: {}\n      value: {}\n      scope: LocalMachine\n",
-                k,
-                yaml_quote(k),
-                yaml_quote(&v.value),
-            ));
-        }
-    }
-    format!(
-        "# Generated by Envarly\n# Requires: DSC v3 (https://github.com/PowerShell/DSC)\n{}$schema: https://aka.ms/DSC/2023/08/config\nresources:\n{}",
-        type_limitation_notice(snapshot, scope, "# "),
-        resources,
-    )
-}
-
-// ── Ansible ───────────────────────────────────────────────────────────────
-
-/// Export snapshot as an Ansible playbook (.yml).
-pub fn to_ansible(snapshot: &EnvSnapshot, scope: ExportScope) -> String {
-    let mut tasks = String::new();
-    if scope.includes_user() {
-        let mut pairs: Vec<_> = snapshot.user.iter().collect();
-        pairs.sort_by_key(|(k, _)| k.to_lowercase());
-        for (k, v) in pairs {
-            tasks.push_str(&format!(
-                "    - name: Set {} (User)\n      ansible.windows.win_environment:\n        name: {}\n        value: {}\n        level: user\n        state: present\n\n",
-                k,
-                yaml_quote(k),
-                yaml_quote(&v.value),
-            ));
-        }
-    }
-    if scope.includes_system() {
-        let mut pairs: Vec<_> = snapshot.system.iter().collect();
-        pairs.sort_by_key(|(k, _)| k.to_lowercase());
-        for (k, v) in pairs {
-            tasks.push_str(&format!(
-                "    - name: Set {} (System)\n      ansible.windows.win_environment:\n        name: {}\n        value: {}\n        level: machine\n        state: present\n\n",
-                k,
-                yaml_quote(k),
-                yaml_quote(&v.value),
-            ));
-        }
-    }
-    format!(
-        "# Generated by Envarly\n{}---\n- name: Configure environment variables\n  hosts: windows\n  gather_facts: false\n  tasks:\n{}\n",
-        type_limitation_notice(snapshot, scope, "# "),
-        tasks.trim_end_matches('\n'),
-    )
-}
-
-// ── Helpers ───────────────────────────────────────────────────────────────
-
-/// Single-quote a string for PowerShell. Single quotes prevent variable expansion.
-/// Internal single quotes are escaped by doubling: ' → ''
-fn ps1_quote(s: &str) -> String {
-    format!("'{}'", s.replace('\'', "''"))
-}
-
-/// Double-quote a string for YAML. Escapes backslashes, double quotes, and control chars.
-fn yaml_quote(s: &str) -> String {
-    let escaped = s
-        .replace('\\', "\\\\")
-        .replace('"', "\\\"")
-        .replace('\n', "\\n")
-        .replace('\r', "\\r")
-        .replace('\t', "\\t");
-    format!("\"{}\"", escaped)
-}
-
-/// Sanitised identifier for a DSC v2 resource instance (alphanumeric + underscore, scope-prefixed).
-fn dsc_id(name: &str, scope: &str) -> String {
-    let sanitised: String = name
-        .chars()
-        .map(|c| if c.is_alphanumeric() { c } else { '_' })
-        .collect();
-    format!("{}_{}", scope, sanitised)
-}
-
-fn reg_escape(s: &str) -> String {
-    s.replace('\\', "\\\\").replace('"', "\\\"")
-}
-
-fn reg_entry(name: &str, value: &EnvValue) -> String {
-    match value.kind {
-        Some(EnvValueKind::ExpandString) => {
-            let bytes = value
-                .value
-                .encode_utf16()
-                .chain(std::iter::once(0))
-                .flat_map(u16::to_le_bytes)
-                .map(|byte| format!("{byte:02x}"))
-                .collect::<Vec<_>>()
-                .join(",");
-            format!("\"{}\"=hex(2):{bytes}\r\n", reg_escape(name))
-        }
-        _ => format!(
-            "\"{}\"=\"{}\"\r\n",
-            reg_escape(name),
-            reg_escape(&value.value)
-        ),
-    }
-}
-
-fn ps1_kind(kind: Option<EnvValueKind>) -> &'static str {
-    match kind {
-        Some(EnvValueKind::ExpandString) => "[Microsoft.Win32.RegistryValueKind]::ExpandString",
-        _ => "[Microsoft.Win32.RegistryValueKind]::String",
-    }
-}
-
-fn type_limitation_notice(snapshot: &EnvSnapshot, scope: ExportScope, prefix: &str) -> String {
-    let includes_expandable = (scope.includes_user()
-        && snapshot
-            .user
-            .values()
-            .any(|value| value.kind == Some(EnvValueKind::ExpandString)))
-        || (scope.includes_system()
-            && snapshot
-                .system
-                .values()
-                .any(|value| value.kind == Some(EnvValueKind::ExpandString)));
-    if includes_expandable {
-        format!(
-            "{prefix}Warning: this format cannot preserve REG_EXPAND_SZ separately from REG_SZ.\n"
-        )
-    } else {
-        String::new()
-    }
-}
-
-fn reg_unescape(s: &str) -> String {
-    let mut out = String::with_capacity(s.len());
-    let mut chars = s.chars().peekable();
-    while let Some(c) = chars.next() {
-        if c == '\\' {
-            match chars.next() {
-                Some('\\') => out.push('\\'),
-                Some('"') => out.push('"'),
-                Some('n') => out.push('\n'),
-                Some(other) => {
-                    out.push('\\');
-                    out.push(other);
-                }
-                None => out.push('\\'),
-            }
-        } else {
-            out.push(c);
-        }
-    }
-    out
-}
-
-/// Parse JSON export back into a snapshot. Does NOT write to the registry.
-pub fn parse_json(content: &str) -> Result<EnvSnapshot, EnvarlyError> {
-    let v: serde_json::Value = serde_json::from_str(content)
-        .map_err(|e| EnvarlyError::InvalidInput(format!("JSON parse error: {e}")))?;
-
-    let as_map = |key: &str| -> HashMap<String, EnvValue> {
-        v.get(key)
-            .and_then(|o| o.as_object())
-            .map(|m| {
-                m.iter()
-                    .filter_map(|(k, vv)| {
-                        if let Some(value) = vv.as_str() {
-                            return Some((
-                                k.clone(),
-                                EnvValue {
-                                    value: value.to_owned(),
-                                    kind: None,
-                                },
-                            ));
-                        }
-                        serde_json::from_value(vv.clone())
-                            .ok()
-                            .map(|value| (k.clone(), value))
-                    })
-                    .collect()
-            })
-            .unwrap_or_default()
-    };
-
-    Ok(EnvSnapshot {
-        user: as_map("user"),
-        system: as_map("system"),
-    })
-}
-
-/// Parse a .reg file back into a snapshot. Does NOT write to the registry.
-pub fn parse_reg(content: &str) -> Result<EnvSnapshot, EnvarlyError> {
-    let mut user: HashMap<String, EnvValue> = HashMap::new();
-    let mut system: HashMap<String, EnvValue> = HashMap::new();
-    let mut current: Option<bool> = None; // Some(true) = user env, Some(false) = system env
-
-    for raw_line in reg_logical_lines(content) {
-        let line = raw_line
-            .trim_start_matches('\u{feff}')
-            .trim_end_matches('\r')
-            .trim();
-        if line.is_empty() || line.starts_with(';') || line.starts_with("Windows Registry Editor") {
-            continue;
-        }
-        if line.starts_with('[') {
-            let lower = line.to_lowercase();
-            if lower.contains("hkey_current_user\\environment") {
-                current = Some(true);
-            } else if lower.contains("session manager\\environment") {
-                current = Some(false);
-            } else {
-                current = None;
-            }
-            continue;
-        }
-        if let Some(is_user) = current {
-            if let Some((name, value)) = parse_reg_entry(line) {
-                if is_user {
-                    user.insert(name, value);
-                } else {
-                    system.insert(name, value);
-                }
-            }
-        }
-    }
-
-    Ok(EnvSnapshot { user, system })
-}
-
-fn reg_logical_lines(content: &str) -> Vec<String> {
-    let mut lines = Vec::new();
-    let mut current = String::new();
-    for raw in content.lines() {
-        let line = raw
-            .trim_start_matches('\u{feff}')
-            .trim_end_matches('\r')
-            .trim();
-        let continues = line.ends_with('\\');
-        current.push_str(line.trim_end_matches('\\').trim_end());
-        if !continues {
-            lines.push(std::mem::take(&mut current));
-        }
-    }
-    if !current.is_empty() {
-        lines.push(current);
-    }
-    lines
-}
-
-/// Parse a single `"name"="value"` registry entry line.
-fn parse_reg_entry(line: &str) -> Option<(String, EnvValue)> {
-    if !line.starts_with('"') {
-        return None; // skip non-string values (dword, binary, etc.)
-    }
-    let after_open = &line[1..]; // skip leading "
-    let (name_raw, after_name) = split_at_unescaped_quote(after_open)?;
-    if let Some(after_eq) = after_name.strip_prefix("=\"") {
-        let (value_raw, _) = split_at_unescaped_quote(after_eq)?;
-        return Some((
-            reg_unescape(name_raw),
-            EnvValue::typed(reg_unescape(value_raw), EnvValueKind::String),
-        ));
-    }
-    if after_name.len() < "=hex(2):".len()
-        || !after_name[.."=hex(2):".len()].eq_ignore_ascii_case("=hex(2):")
-    {
-        return None;
-    }
-    let bytes = &after_name["=hex(2):".len()..];
-    let raw = bytes
-        .split(',')
-        .map(|part| u8::from_str_radix(part.trim(), 16).ok())
-        .collect::<Option<Vec<_>>>()?;
-    if raw.len() % 2 != 0 {
-        return None;
-    }
-    let mut units: Vec<u16> = raw
-        .chunks_exact(2)
-        .map(|pair| u16::from_le_bytes([pair[0], pair[1]]))
-        .collect();
-    if units.last() == Some(&0) {
-        units.pop();
-    }
-    Some((
-        reg_unescape(name_raw),
-        EnvValue::typed(String::from_utf16(&units).ok()?, EnvValueKind::ExpandString),
-    ))
-}
-
-/// Returns (before_quote, after_quote) split at the first unescaped `"`.
-fn split_at_unescaped_quote(s: &str) -> Option<(&str, &str)> {
-    let mut escaped = false;
-    for (i, c) in s.char_indices() {
-        if escaped {
-            escaped = false;
-        } else if c == '\\' {
-            escaped = true;
-        } else if c == '"' {
-            return Some((&s[..i], &s[i + 1..]));
-        }
-    }
-    None
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn string(value: &str) -> EnvValue {
-        EnvValue::typed(value.to_string(), EnvValueKind::String)
-    }
-
-    fn expandable(value: &str) -> EnvValue {
-        EnvValue::typed(value.to_string(), EnvValueKind::ExpandString)
-    }
-
-    fn fixture() -> EnvSnapshot {
-        EnvSnapshot {
-            user: [
-                (
-                    "PATH".to_string(),
-                    expandable("%USERPROFILE%\\bin;C:\\tools"),
-                ),
-                ("JAVA_HOME".to_string(), string("C:\\jdk21")),
-            ]
-            .into(),
-            system: [
-                ("WINDIR".to_string(), string("C:\\Windows")),
-                ("OS".to_string(), string("Windows_NT")),
-            ]
-            .into(),
-        }
-    }
-
-    // ── JSON ──────────────────────────────────────────────────────────────
-
-    #[test]
-    fn json_roundtrip_all() {
-        let snap = fixture();
-        let json = to_json(&snap, ExportScope::All);
-        assert!(json.contains("\"version\": 2"));
-        let parsed = parse_json(&json).unwrap();
-        assert_eq!(parsed.user, snap.user);
-        assert_eq!(parsed.system, snap.system);
-    }
-
-    #[test]
-    fn json_roundtrip_user_only() {
-        let snap = fixture();
-        let parsed = parse_json(&to_json(&snap, ExportScope::User)).unwrap();
-        assert_eq!(parsed.user, snap.user);
-        assert!(parsed.system.is_empty());
-    }
-
-    #[test]
-    fn json_roundtrip_system_only() {
-        let snap = fixture();
-        let parsed = parse_json(&to_json(&snap, ExportScope::System)).unwrap();
-        assert!(parsed.user.is_empty());
-        assert_eq!(parsed.system, snap.system);
-    }
-
-    #[test]
-    fn json_invalid_returns_error() {
-        assert!(parse_json("not json at all").is_err());
-    }
-
-    #[test]
-    fn json_missing_sections_gives_empty_maps() {
-        let parsed = parse_json("{}").unwrap();
-        assert!(parsed.user.is_empty());
-        assert!(parsed.system.is_empty());
-    }
-
-    #[test]
-    fn json_legacy_values_remain_unresolved() {
-        let parsed = parse_json(r#"{"user":{"TEMP":"%USERPROFILE%\\Temp"}}"#).unwrap();
-        assert_eq!(parsed.user["TEMP"].value, "%USERPROFILE%\\Temp");
-        assert_eq!(parsed.user["TEMP"].kind, None);
-    }
-
-    // ── REG ───────────────────────────────────────────────────────────────
-
-    #[test]
-    fn reg_roundtrip_all() {
-        let snap = fixture();
-        let parsed = parse_reg(&to_reg(&snap, ExportScope::All)).unwrap();
-        assert_eq!(parsed.user, snap.user);
-        assert_eq!(parsed.system, snap.system);
-    }
-
-    #[test]
-    fn reg_roundtrip_user_only() {
-        let snap = fixture();
-        let parsed = parse_reg(&to_reg(&snap, ExportScope::User)).unwrap();
-        assert_eq!(parsed.user, snap.user);
-        assert!(parsed.system.is_empty());
-    }
-
-    #[test]
-    fn reg_escape_backslash_roundtrip() {
-        let snap = EnvSnapshot {
-            user: [("X".to_string(), string("C:\\foo\\bar"))].into(),
-            system: Default::default(),
-        };
-        let parsed = parse_reg(&to_reg(&snap, ExportScope::User)).unwrap();
-        assert_eq!(parsed.user["X"].value, "C:\\foo\\bar");
-    }
-
-    #[test]
-    fn reg_escape_quote_roundtrip() {
-        let snap = EnvSnapshot {
-            user: [("X".to_string(), string(r#"say "hello""#))].into(),
-            system: Default::default(),
-        };
-        let parsed = parse_reg(&to_reg(&snap, ExportScope::User)).unwrap();
-        assert_eq!(parsed.user["X"].value, r#"say "hello""#);
-    }
-
-    #[test]
-    fn reg_ignores_unrelated_sections() {
-        let content = "Windows Registry Editor Version 5.00\r\n\
-                       \r\n[HKEY_LOCAL_MACHINE\\SOFTWARE\\SomeApp]\r\n\
-                       \"key\"=\"value\"\r\n";
-        let parsed = parse_reg(content).unwrap();
-        assert!(parsed.user.is_empty());
-        assert!(parsed.system.is_empty());
-    }
-
-    #[test]
-    fn reg_skips_non_string_values() {
-        let content = "Windows Registry Editor Version 5.00\r\n\
-                       \r\n[HKEY_CURRENT_USER\\Environment]\r\n\
-                       \"PATH\"=\"C:\\\\bin\"\r\n\
-                       \"DWORD_VAL\"=dword:00000001\r\n";
-        let parsed = parse_reg(content).unwrap();
-        assert_eq!(parsed.user["PATH"].value, "C:\\bin");
-        assert!(!parsed.user.contains_key("DWORD_VAL"));
-    }
-
-    #[test]
-    fn reg_parses_folded_expand_string() {
-        let content = "Windows Registry Editor Version 5.00\r\n\
-            \r\n[HKEY_CURRENT_USER\\Environment]\r\n\
-            \"TEMP\"=hex(2):25,00,55,00,53,00,45,00,52,00,50,00,52,00,4f,00,46,00,49,00,\\\r\n\
-              4c,00,45,00,25,00,5c,00,54,00,65,00,6d,00,70,00,00,00\r\n";
-        let parsed = parse_reg(content).unwrap();
-        assert_eq!(parsed.user["TEMP"].value, "%USERPROFILE%\\Temp");
-        assert_eq!(parsed.user["TEMP"].kind, Some(EnvValueKind::ExpandString));
-    }
-
-    // ── IaC helpers ───────────────────────────────────────────────────────
-
-    #[test]
-    fn ps1_quote_simple() {
-        assert_eq!(ps1_quote("hello"), "'hello'");
-    }
-
-    #[test]
-    fn ps1_quote_with_single_quote() {
-        assert_eq!(ps1_quote("it's"), "'it''s'");
-    }
-
-    #[test]
-    fn ps1_quote_backslash_unchanged() {
-        // Backslashes don't need escaping in single-quoted PS strings
-        assert_eq!(ps1_quote("C:\\foo"), "'C:\\foo'");
-    }
-
-    #[test]
-    fn yaml_quote_simple() {
-        assert_eq!(yaml_quote("hello"), "\"hello\"");
-    }
-
-    #[test]
-    fn yaml_quote_backslash() {
-        assert_eq!(yaml_quote("C:\\foo"), "\"C:\\\\foo\"");
-    }
-
-    #[test]
-    fn yaml_quote_double_quote() {
-        assert_eq!(yaml_quote(r#"say "hi""#), r#""say \"hi\"""#);
-    }
-
-    #[test]
-    fn dsc_id_sanitises() {
-        assert_eq!(dsc_id("MY-VAR", "User"), "User_MY_VAR");
-    }
-
-    // ── IaC generators ────────────────────────────────────────────────────
-
-    #[test]
-    fn ps1_contains_set_env_calls() {
-        let snap = fixture();
-        let out = to_ps1(&snap, ExportScope::All);
-        assert!(out.contains("$userEnvironment.SetValue('PATH'"));
-        assert!(out.contains("RegistryValueKind]::ExpandString"));
-        assert!(out.contains("$systemEnvironment.SetValue('WINDIR'"));
-    }
-
-    #[test]
-    fn ps1_user_scope_omits_system() {
-        let snap = fixture();
-        let out = to_ps1(&snap, ExportScope::User);
-        assert!(out.contains("$userEnvironment.SetValue"));
-        assert!(!out.contains("$systemEnvironment.SetValue('WINDIR'"));
-    }
-
-    #[test]
-    fn dsc_v2_contains_configuration_block() {
-        let snap = fixture();
-        let out = to_dsc_v2(&snap, ExportScope::All);
-        assert!(out.contains("Configuration EnvarlyConfig"));
-        assert!(out.contains("Import-DscResource -ModuleName PSDscResources"));
-        assert!(out.contains("Environment User_PATH"));
-        assert!(out.contains("Target = @('User')"));
-        assert!(out.contains("Target = @('Machine')"));
-    }
-
-    #[test]
-    fn dsc_v3_contains_schema_and_resources() {
-        let snap = fixture();
-        let out = to_dsc_v3(&snap, ExportScope::All);
-        assert!(out.contains("$schema: https://aka.ms/DSC/2023/08/config"));
-        assert!(out.contains("type: Microsoft.Windows/Environment"));
-        assert!(out.contains("scope: CurrentUser"));
-        assert!(out.contains("scope: LocalMachine"));
-    }
-
-    #[test]
-    fn ansible_contains_win_environment() {
-        let snap = fixture();
-        let out = to_ansible(&snap, ExportScope::All);
-        assert!(out.contains("ansible.windows.win_environment:"));
-        assert!(out.contains("level: user"));
-        assert!(out.contains("level: machine"));
-        assert!(out.contains("state: present"));
-    }
-
-    #[test]
-    fn ansible_system_scope_omits_user() {
-        let snap = fixture();
-        let out = to_ansible(&snap, ExportScope::System);
-        assert!(!out.contains("level: user"));
-        assert!(out.contains("level: machine"));
-    }
-}
