+++ src-tauri/src/export/ansible.rs
@@ -1,0 +1,86 @@
+use super::{type_limitation_notice, yaml_quote, ExportScope};
+use crate::env_store::EnvSnapshot;
+
+/// Export snapshot as an Ansible playbook (.yml).
+pub fn to_ansible(snapshot: &EnvSnapshot, scope: ExportScope) -> String {
+    let mut tasks = String::new();
+    if scope.includes_user() {
+        let mut pairs: Vec<_> = snapshot.user.iter().collect();
+        pairs.sort_by_key(|(k, _)| k.to_lowercase());
+        for (k, v) in pairs {
+            tasks.push_str(&format!(
+                "    - name: Set {} (User)\n      ansible.windows.win_environment:\n        name: {}\n        value: {}\n        level: user\n        state: present\n\n",
+                k,
+                yaml_quote(k),
+                yaml_quote(&v.value),
+            ));
+        }
+    }
+    if scope.includes_system() {
+        let mut pairs: Vec<_> = snapshot.system.iter().collect();
+        pairs.sort_by_key(|(k, _)| k.to_lowercase());
+        for (k, v) in pairs {
+            tasks.push_str(&format!(
+                "    - name: Set {} (System)\n      ansible.windows.win_environment:\n        name: {}\n        value: {}\n        level: machine\n        state: present\n\n",
+                k,
+                yaml_quote(k),
+                yaml_quote(&v.value),
+            ));
+        }
+    }
+    format!(
+        "# Generated by Envarly\n{}---\n- name: Configure environment variables\n  hosts: windows\n  gather_facts: false\n  tasks:\n{}\n",
+        type_limitation_notice(snapshot, scope, "# "),
+        tasks.trim_end_matches('\n'),
+    )
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::env_store::{EnvValue, EnvValueKind};
+
+    fn string(value: &str) -> EnvValue {
+        EnvValue::typed(value.to_string(), EnvValueKind::String)
+    }
+
+    fn expandable(value: &str) -> EnvValue {
+        EnvValue::typed(value.to_string(), EnvValueKind::ExpandString)
+    }
+
+    fn fixture() -> EnvSnapshot {
+        EnvSnapshot {
+            user: [
+                (
+                    "PATH".to_string(),
+                    expandable("%USERPROFILE%\\bin;C:\\tools"),
+                ),
+                ("JAVA_HOME".to_string(), string("C:\\jdk21")),
+            ]
+            .into(),
+            system: [
+                ("WINDIR".to_string(), string("C:\\Windows")),
+                ("OS".to_string(), string("Windows_NT")),
+            ]
+            .into(),
+        }
+    }
+
+    #[test]
+    fn ansible_contains_win_environment() {
+        let snap = fixture();
+        let out = to_ansible(&snap, ExportScope::All);
+        assert!(out.contains("ansible.windows.win_environment:"));
+        assert!(out.contains("level: user"));
+        assert!(out.contains("level: machine"));
+        assert!(out.contains("state: present"));
+    }
+
+    #[test]
+    fn ansible_system_scope_omits_user() {
+        let snap = fixture();
+        let out = to_ansible(&snap, ExportScope::System);
+        assert!(!out.contains("level: user"));
+        assert!(out.contains("level: machine"));
+    }
+}
