+++ src-tauri/src/export/dsc.rs
@@ -1,0 +1,137 @@
+use super::{ps1_quote, type_limitation_notice, yaml_quote, ExportScope};
+use crate::env_store::EnvSnapshot;
+
+/// Sanitised identifier for a DSC v2 resource instance (alphanumeric + underscore, scope-prefixed).
+fn dsc_id(name: &str, scope: &str) -> String {
+    let sanitised: String = name
+        .chars()
+        .map(|c| if c.is_alphanumeric() { c } else { '_' })
+        .collect();
+    format!("{}_{}", scope, sanitised)
+}
+
+/// Export snapshot as a PowerShell DSC v2 Configuration (.ps1).
+pub fn to_dsc_v2(snapshot: &EnvSnapshot, scope: ExportScope) -> String {
+    let mut resources = String::new();
+    if scope.includes_user() {
+        let mut pairs: Vec<_> = snapshot.user.iter().collect();
+        pairs.sort_by_key(|(k, _)| k.to_lowercase());
+        for (k, v) in pairs {
+            resources.push_str(&format!(
+                "        Environment {} {{\n            Name   = {}\n            Value  = {}\n            Target = @('User')\n        }}\n",
+                dsc_id(k, "User"),
+                ps1_quote(k),
+                ps1_quote(&v.value),
+            ));
+        }
+    }
+    if scope.includes_system() {
+        let mut pairs: Vec<_> = snapshot.system.iter().collect();
+        pairs.sort_by_key(|(k, _)| k.to_lowercase());
+        for (k, v) in pairs {
+            resources.push_str(&format!(
+                "        Environment {} {{\n            Name   = {}\n            Value  = {}\n            Target = @('Machine')\n        }}\n",
+                dsc_id(k, "Machine"),
+                ps1_quote(k),
+                ps1_quote(&v.value),
+            ));
+        }
+    }
+    format!(
+        "# Generated by Envarly\n{}Configuration EnvarlyConfig {{\n    Import-DscResource -ModuleName PSDscResources\n\n    Node localhost {{\n{}    }}\n}}\nEnvarlyConfig\n",
+        type_limitation_notice(snapshot, scope, "# "),
+        resources,
+    )
+}
+
+/// Export snapshot as a DSC v3 config document (.dsc.yaml).
+pub fn to_dsc_v3(snapshot: &EnvSnapshot, scope: ExportScope) -> String {
+    let mut resources = String::new();
+    if scope.includes_user() {
+        let mut pairs: Vec<_> = snapshot.user.iter().collect();
+        pairs.sort_by_key(|(k, _)| k.to_lowercase());
+        for (k, v) in pairs {
+            resources.push_str(&format!(
+                "  - name: Set {} (User)\n    type: Microsoft.Windows/Environment\n    properties:\n      name: {}\n      value: {}\n      scope: CurrentUser\n",
+                k,
+                yaml_quote(k),
+                yaml_quote(&v.value),
+            ));
+        }
+    }
+    if scope.includes_system() {
+        let mut pairs: Vec<_> = snapshot.system.iter().collect();
+        pairs.sort_by_key(|(k, _)| k.to_lowercase());
+        for (k, v) in pairs {
+            resources.push_str(&format!(
+                "  - name: Set {} (System)\n    type: Microsoft.Windows/Environment\n    properties:\n      name: {}\n      value: {}\n      scope: LocalMachine\n",
+                k,
+                yaml_quote(k),
+                yaml_quote(&v.value),
+            ));
+        }
+    }
+    format!(
+        "# Generated by Envarly\n# Requires: DSC v3 (https://github.com/PowerShell/DSC)\n{}$schema: https://aka.ms/DSC/2023/08/config\nresources:\n{}",
+        type_limitation_notice(snapshot, scope, "# "),
+        resources,
+    )
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::env_store::{EnvValue, EnvValueKind};
+
+    fn string(value: &str) -> EnvValue {
+        EnvValue::typed(value.to_string(), EnvValueKind::String)
+    }
+
+    fn expandable(value: &str) -> EnvValue {
+        EnvValue::typed(value.to_string(), EnvValueKind::ExpandString)
+    }
+
+    fn fixture() -> EnvSnapshot {
+        EnvSnapshot {
+            user: [
+                (
+                    "PATH".to_string(),
+                    expandable("%USERPROFILE%\\bin;C:\\tools"),
+                ),
+                ("JAVA_HOME".to_string(), string("C:\\jdk21")),
+            ]
+            .into(),
+            system: [
+                ("WINDIR".to_string(), string("C:\\Windows")),
+                ("OS".to_string(), string("Windows_NT")),
+            ]
+            .into(),
+        }
+    }
+
+    #[test]
+    fn dsc_id_sanitises() {
+        assert_eq!(dsc_id("MY-VAR", "User"), "User_MY_VAR");
+    }
+
+    #[test]
+    fn dsc_v2_contains_configuration_block() {
+        let snap = fixture();
+        let out = to_dsc_v2(&snap, ExportScope::All);
+        assert!(out.contains("Configuration EnvarlyConfig"));
+        assert!(out.contains("Import-DscResource -ModuleName PSDscResources"));
+        assert!(out.contains("Environment User_PATH"));
+        assert!(out.contains("Target = @('User')"));
+        assert!(out.contains("Target = @('Machine')"));
+    }
+
+    #[test]
+    fn dsc_v3_contains_schema_and_resources() {
+        let snap = fixture();
+        let out = to_dsc_v3(&snap, ExportScope::All);
+        assert!(out.contains("$schema: https://aka.ms/DSC/2023/08/config"));
+        assert!(out.contains("type: Microsoft.Windows/Environment"));
+        assert!(out.contains("scope: CurrentUser"));
+        assert!(out.contains("scope: LocalMachine"));
+    }
+}
