+++ src-tauri/src/export/powershell.rs
@@ -1,0 +1,98 @@
+use super::{ps1_quote, ExportScope};
+use crate::env_store::{EnvSnapshot, EnvValueKind};
+
+/// Export snapshot as a PowerShell script that preserves registry value types.
+pub fn to_ps1(snapshot: &EnvSnapshot, scope: ExportScope) -> String {
+    let mut out = String::from("# Generated by Envarly\n");
+    if scope.includes_user() && !snapshot.user.is_empty() {
+        out.push_str(
+            "$userEnvironment = [Microsoft.Win32.Registry]::CurrentUser.OpenSubKey('Environment', $true)\n",
+        );
+        out.push_str("\n# User environment variables\n");
+        let mut pairs: Vec<_> = snapshot.user.iter().collect();
+        pairs.sort_by_key(|(k, _)| k.to_lowercase());
+        for (k, v) in pairs {
+            out.push_str(&format!(
+                "$userEnvironment.SetValue({}, {}, {})\n",
+                ps1_quote(k),
+                ps1_quote(&v.value),
+                ps1_kind(v.kind),
+            ));
+        }
+        out.push_str("$userEnvironment.Dispose()\n");
+    }
+    if scope.includes_system() && !snapshot.system.is_empty() {
+        out.push_str(
+            "$systemEnvironment = [Microsoft.Win32.Registry]::LocalMachine.OpenSubKey('SYSTEM\\CurrentControlSet\\Control\\Session Manager\\Environment', $true)\n",
+        );
+        out.push_str("\n# System environment variables\n");
+        let mut pairs: Vec<_> = snapshot.system.iter().collect();
+        pairs.sort_by_key(|(k, _)| k.to_lowercase());
+        for (k, v) in pairs {
+            out.push_str(&format!(
+                "$systemEnvironment.SetValue({}, {}, {})\n",
+                ps1_quote(k),
+                ps1_quote(&v.value),
+                ps1_kind(v.kind),
+            ));
+        }
+        out.push_str("$systemEnvironment.Dispose()\n");
+    }
+    out
+}
+
+fn ps1_kind(kind: Option<EnvValueKind>) -> &'static str {
+    match kind {
+        Some(EnvValueKind::ExpandString) => "[Microsoft.Win32.RegistryValueKind]::ExpandString",
+        _ => "[Microsoft.Win32.RegistryValueKind]::String",
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::env_store::EnvValue;
+
+    fn string(value: &str) -> EnvValue {
+        EnvValue::typed(value.to_string(), EnvValueKind::String)
+    }
+
+    fn expandable(value: &str) -> EnvValue {
+        EnvValue::typed(value.to_string(), EnvValueKind::ExpandString)
+    }
+
+    fn fixture() -> EnvSnapshot {
+        EnvSnapshot {
+            user: [
+                (
+                    "PATH".to_string(),
+                    expandable("%USERPROFILE%\\bin;C:\\tools"),
+                ),
+                ("JAVA_HOME".to_string(), string("C:\\jdk21")),
+            ]
+            .into(),
+            system: [
+                ("WINDIR".to_string(), string("C:\\Windows")),
+                ("OS".to_string(), string("Windows_NT")),
+            ]
+            .into(),
+        }
+    }
+
+    #[test]
+    fn ps1_contains_set_env_calls() {
+        let snap = fixture();
+        let out = to_ps1(&snap, ExportScope::All);
+        assert!(out.contains("$userEnvironment.SetValue('PATH'"));
+        assert!(out.contains("RegistryValueKind]::ExpandString"));
+        assert!(out.contains("$systemEnvironment.SetValue('WINDIR'"));
+    }
+
+    #[test]
+    fn ps1_user_scope_omits_system() {
+        let snap = fixture();
+        let out = to_ps1(&snap, ExportScope::User);
+        assert!(out.contains("$userEnvironment.SetValue"));
+        assert!(!out.contains("$systemEnvironment.SetValue('WINDIR'"));
+    }
+}
